chore: moves controls
