+++ DesktopUI2/DesktopUI2/Views/Controls/ReceiveButton.xaml
@@ -1,5 +1,5 @@
 <UserControl
-  x:Class="DesktopUI2.Controls.ReceiveButton"
+  x:Class="DesktopUI2.Views.Controls.ReceiveButton"
   xmlns="https://github.com/avaloniaui"
   xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
   xmlns:assists="clr-namespace:Material.Styles.Assists;assembly=Material.Styles"
@@ -46,10 +46,10 @@
       <ProgressBar
         Width="40"
         Height="40"
-        Classes="Circle"
+        Classes="Circle notransitions"
+        IsEnabled="{Binding !Progress.IsPreviewProgressing, FallbackValue=True}"
         IsIndeterminate="{Binding Progress.IsIndeterminate, FallbackValue=True}"
         IsVisible="{Binding Progress.IsProgressing, FallbackValue=False}"
-        IsEnabled="{Binding !Progress.IsPreviewProgressing, FallbackValue=True}"
         Maximum="{Binding Progress.Max}"
         Value="{Binding Progress.Value}" />
       <m:FloatingButton
@@ -58,8 +58,8 @@
         Background="Transparent"
         Classes="Mini"
         Command="{Binding CancelSendOrReceiveCommand}"
-        IsVisible="{Binding Progress.IsProgressing, FallbackValue=False}"
         IsEnabled="{Binding !Progress.IsPreviewProgressing, FallbackValue=True}"
+        IsVisible="{Binding Progress.IsProgressing, FallbackValue=False}"
         ToolTip.Tip="Cancel">
         <Button.Content>
           <icons:MaterialIcon
