+++ DesktopUI2/DesktopUI2/Views/Controls/ReceiveButton.xaml.cs
@@ -2,11 +2,11 @@ using Avalonia;
 using Avalonia.Controls;
 using Avalonia.Markup.Xaml;
 
-namespace DesktopUI2.Controls
+namespace DesktopUI2.Views.Controls
 {
-  public partial class SendButton : UserControl
+  public partial class ReceiveButton : UserControl
   {
-    public SendButton()
+    public ReceiveButton()
     {
       InitializeComponent();
     }
